+++ evm_loader/faucet/src/server.rs
@@ -7,7 +7,7 @@ use actix_web::{App, HttpResponse, HttpServer, Responder};
 use color_eyre::Result;
 use tracing::{error, info};
 
-use crate::{config, solana, tokens};
+use crate::{config, eth_token, tokens};
 
 /// Starts the server in listening mode.
 pub async fn start(rpc_port: u16, workers: usize) -> Result<()> {
@@ -43,13 +43,13 @@ async fn handle_request_eth_token(body: Bytes) -> impl Responder {
         return HttpResponse::BadRequest();
     }
 
-    let airdrop = serde_json::from_str::<solana::Airdrop>(&input.unwrap());
+    let airdrop = serde_json::from_str::<eth_token::Airdrop>(&input.unwrap());
     if let Err(err) = airdrop {
         error!("BadRequest: {}", err);
         return HttpResponse::BadRequest();
     }
 
-    if let Err(err) = solana::airdrop(airdrop.unwrap()).await {
+    if let Err(err) = eth_token::airdrop(airdrop.unwrap()).await {
         error!("InternalServerError: {}", err);
         return HttpResponse::InternalServerError();
     }
